refactoring files (#1476)
