+++ wondrous-app/components/Common/MoreInfoModal/index.tsx
