+++ wondrous-app/components/Common/index.ts
@@ -1,11 +1,6 @@
 import { Typography } from '@mui/material';
 import styled from 'styled-components';
-import { greyColors } from 'theme/colors';
 import palette from 'theme/palette';
-
-export const Flex = styled.div`
-  flex: 1;
-`;
 
 export const FlexRow = styled.div`
   display: flex;
@@ -16,14 +11,6 @@ export const CenteredFlexRow = styled(FlexRow)`
   align-items: center;
   justify-content: center;
   margin-top: ${(props) => props.marginTop || 0};
-`;
-
-export const NewCanvas = styled.canvas`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 50px;
-  height: 50px;
 `;
 
 export const ErrorText = styled(Typography)`
