+++ wondrous-app/components/Pod/analytics/index.tsx
@@ -22,7 +22,7 @@ import {
 } from 'components/organization/analytics/styles';
 import { PayoutModal } from 'components/organization/analytics/PayoutModal';
 import { SafeImage } from 'components/Common/Image';
-import { DefaultProfilePicture, UserProfilePicture } from 'components/profile/modals/styles';
+import { DefaultProfilePicture, UserProfilePicture } from 'components/Common/MoreInfoModal/styles';
 import DefaultUserImage from 'components/Common/Image/DefaultUserImage';
 import BottomArrowCaret from 'components/Icons/BottomArrowCaret';
 import RightArrowCaret from 'components/Icons/RightArrowCaret';
