+++ wondrous-app/components/Pod/wrapper/HomePageHeader.tsx
@@ -22,7 +22,7 @@ import { SafeImage } from 'components/Common/Image';
 import { RichTextViewer } from 'components/RichText';
 import ChooseEntityToCreate from 'components/CreateEntity';
 import RolePill from 'components/Common/RolePill';
-import MoreInfoModal from 'components/profile/modals';
+import MoreInfoModal from 'components/Common/MoreInfoModal';
 import MembersIcon from 'components/Icons/members';
 import HeaderSocialLinks from 'components/organization/wrapper/HeaderSocialLinks';
 import { Button as PrimaryButton } from 'components/Button';
