--- wondrous-app/components/profile/about/styles.tsx
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-const OrganisationsCardNoLogo = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  max-height: 40px;
-  width: 40px;
-`;
-
-export default OrganisationsCardNoLogo;
